Iron out conversion issues

+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/CodeActions/RazorCodeActionRunner.ts
@@ -27,30 +27,48 @@ export class RazorCodeActionRunner {
 
     private async runCodeAction(action: string, args: string[]): Promise<boolean | string | {}> {
         const response: CodeActionComputationResponse = await this.serverClient.sendRequest('razor/codeActionComputation', {Action: action, Arguments: args});
-        this.logger.logMessage(` ${JSON.stringify(response)}`);
+        this.logger.logMessage(`Received computed workspace edit ${JSON.stringify(response)}`);
 
         const workspaceEdit = new vscode.WorkspaceEdit();
-        for (const documentChange of response.edit.documentChanges) {
-            if (documentChange.kind === 'create') {
-                workspaceEdit.createFile(vscode.Uri.parse(documentChange.uri), documentChange.options);
-            } else if (documentChange.kind === 'rename') {
-                workspaceEdit.renameFile(vscode.Uri.parse(documentChange.oldUri), vscode.Uri.parse(documentChange.newUri), documentChange.options);
-            } else if (documentChange.kind === 'delete') {
-                workspaceEdit.deleteFile(vscode.Uri.parse(documentChange.uri), documentChange.options);
+        try {
+            if (Array.isArray(response.edit.documentChanges)) {
+                for (const documentChange of response.edit.documentChanges) {
+                    if (documentChange.kind === 'create') {
+                        workspaceEdit.createFile(vscode.Uri.parse(documentChange.uri), documentChange.options);
+                    } else if (documentChange.kind === 'rename') {
+                        workspaceEdit.renameFile(vscode.Uri.parse(documentChange.oldUri), vscode.Uri.parse(documentChange.newUri), documentChange.options);
+                    } else if (documentChange.kind === 'delete') {
+                        workspaceEdit.deleteFile(vscode.Uri.parse(documentChange.uri), documentChange.options);
+                    }
+                }
             }
+
+            if (response.edit.changes !== undefined) {
+                for (const uri in response.edit.changes) {
+                    if (!response.edit.changes.hasOwnProperty(uri)) {
+                        continue;
+                    }
+                    const changes = response.edit.changes[uri];
+                    for (const change of changes) {
+                        const range = new vscode.Range(
+                            change.range.start.line,
+                            change.range.start.character,
+                            change.range.end.line,
+                            change.range.end.character);
+                        workspaceEdit.replace(vscode.Uri.parse(uri), range, change.newText);
+                    }
+                }
+            }
+        } catch (e) {
+            this.logger.logAlways(`caught in run code action: ${e}`);
+            return Promise.resolve(false);
         }
 
-        for (const [uri, change] of response.edit.changes.entries()) {
-            const range = new vscode.Range(
-                change.range.start.line,
-                change.range.start.character,
-                change.range.end.line,
-                change.range.end.character);
-            workspaceEdit.replace(vscode.Uri.parse(uri), range, change.newText);
-        }
+        const result = await vscode.workspace.applyEdit(workspaceEdit);
+        this.logger.logAlways(` status ${result}`);
 
-        await vscode.workspace.applyEdit(workspaceEdit);
         return true;
+
         //     // Unfortunately, the textEditor.Close() API has been deprecated
         //     // and replaced with a command that can only close the active editor.
         //     // If files were renamed that weren't the active editor, their tabs will
